+++ openbbs/shell.py
@@ -1,62 +1,15 @@
 """Interactive shell for communication between the client and BBS."""
 
 import logging
-import textwrap
 import time
 
 from openbbs import __version__
 from openbbs.command import CommandInterpreter
-
-
-def box_boards(boards):
-    """Text formatter for the board listing."""
-    string = "+=============================================================="\
-             "================+\r\n|                                BOARD LIS"\
-             "TING                                 |\r\n+===================="\
-             "==========================================================+\r\n"
-    for title, description in [board.split(":") for board in boards]:
-        string += "| %-18s | %55s |\r\n+====================================="\
-                  "=========================================+\r\n" % \
-                  (title, description)
-    return string
-
-
-def box_posts(posts):
-    """Text formatter for a listing of threads in a board."""
-    string = "+=============================================================="\
-             "================+\r\n|                                THREAD LI"\
-             "STING                                |\r\n+===================="\
-             "==========================================================+\r\n"
-    for post_id, pub_time, poster, subject, body in posts:
-        printable_time = time.strftime("%m/%d/%y %H:%M:%S",
-                                       time.localtime(pub_time))
-        string += "| #%-6d|%.17s| %-19s| %-29.29s|\r\n+======================"\
-        "========================================================+" % \
-        (post_id, printable_time, poster, subject.strip())
-    return string
-
-
-def box_thread(posts):
-    """Returns a table-formatted version of the thread being viewed."""
-    title = posts[0][3]
-    string = "+=============================================================="\
-             "================+\r\n|" + ((78 - len(title)) // 2) * " " + title\
-             + ((78 - len(title)) // 2) * " " + "|\r\n+======================"\
-             "========================================================+\r\n"
-    for post in posts:
-        post_time = time.ctime(post[1])
-        string += "| #%-9d | %-26s posted on %-26s |\r\n+===================="\
-                  "=========================================================="\
-                  "+\r\n" % (post[0], post[2], post_time)
-        for line in textwrap.wrap(post[4], width=76):
-            string += "| %-76s |\r\n" % line
-        string += "+========================================================="\
-                  "=====================+\r\n"
-    return string
+from openbbs.formatters import (box_boards, box_posts, box_thread)
 
 
 def handle_bogus_input(user, parameters):
-    """[Document me!]"""
+    """Generic handler for invalid commands."""
     user.send("Unknown command: \"%s\"" % parameters[0])
 
 
@@ -116,16 +69,18 @@ def change_thread(user, parameters):
     """Changes the user's current thread, if valid."""
     if user.current_board == "main":
         user.send("There are no threads here.")
-    elif len(parameters) > 1:
-        user.current_thread = parameters[1]
     else:
-        user.send("Leave empty to return to the thread listing.\r\nTHREAD "
-                  "NUMBER: ", end="")
-        thread = user.receive()
+        if len(parameters) > 1:
+            thread = parameters[1]
+        else:
+            user.send("Leave empty to return to the thread listing.\r\nTHREAD "
+                      "NUMBER: ", end="")
+            thread = user.receive()
+
         if thread == "":
             user.current_thread = None
             user.send(box_posts(user.database.get_posts(user.current_board)))
-            user.send("Successfully returned to the %s home." % current_board)
+            user.send("Returned to the %s home." % user.current_board)
         else:
             posts = user.database.get_posts(user.current_board, thread)
             if posts:
@@ -133,11 +88,11 @@ def change_thread(user, parameters):
                 user.send(box_thread(posts))
                 user.send("Current thread changed to %s." % thread)
             else:
-                current_thread = None
+                user.current_thread = None
                 user.send("Thread %s does not exist." % thread)
 
 
-def get_inbox(user, parameters):
+def get_inbox(user, _):
     """Sends the user their inbox, provided they are not anonymous."""
     if user.status != "coward":
         messages = user.database.get_pms(user.name)
@@ -167,7 +122,8 @@ def make_post(user, _):
             subject = user.receive()
             user.send("BODY: ", end="")
             body = user.receive()
-            database.make_post(user.name, subject, body, user.current_board)
+            user.database.make_post(user.name, subject, body,
+                                    user.current_board)
             user.send("Successfully posted.")
 
 
@@ -193,7 +149,7 @@ def send_message(user, parameters):
             user.send("RECEIVER: ", end="")
             receiver = user.receive().lower()
         if len(parameters) > 2:
-            message = command[2]
+            message = parameters[2]
         else:
             user.send("MESSAGE: ", end="")
             message = user.receive()
@@ -212,7 +168,7 @@ def delete_post(user, parameters):
             target = int(parameters[1])
         else:
             user.send("POST ID: ", end="")
-            target = user.receive()
+            target = int(user.receive())
         user.database.delete_post(target)
         user.send("Post %d successfully deleted." % target)
     else:
@@ -228,7 +184,7 @@ def ban_user(user, parameters):
             user.send("USER: ", end="")
             target = user.receive()
         if len(parameters) > 2:
-            reason = command[2]
+            reason = parameters[2]
         else:
             user.send("REASON: ", end="")
             reason = user.receive()