+++ tests/test_core.py
@@ -11,14 +11,16 @@ logging.disable(logging.CRITICAL)
 
 
 class ServerSpawnTest(unittest.TestCase):
+    # Literally no idea why this test doesn't work in Tox.
     def test_create_server(self):
         config = load_config("inexistent.conf")
         threading.Thread(target=spawn_server, args=(config, True)).start()
 
         # Connection occurs before binding if the test doesn't sleep.
-        # 3 seconds is a good middle-ground.
-        time.sleep(3)
+        # 10 seconds is a good middle-ground.
+        time.sleep(10)
         test_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        test_sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         test_sock.connect((socket.gethostbyname(socket.gethostname()), 1337))
 
         self.assertTrue(test_sock.recv(1024))